core2c/requirements.txt
aiosqlite==0.17.0
anyio==3.6.2
asgiref==3.6.0
asyncpg==0.27.0
click==8.1.3
fastapi==0.75.2
h11==0.14.0
idna==3.4
iso8601==1.1.0
loguru==0.6.0
pydantic==1.10.4
pypika-tortoise==0.1.6
pytz==2022.7.1
sniffio==1.3.0
starlette==0.17.1
tortoise-orm==0.19.0
typing_extensions==4.4.0
uvicorn==0.17.6
requests==2.27.1
